+++ src/main/java/com/example/demo/services/InstockDetailServiceDao.java
@@ -11,7 +11,7 @@ import java.util.Map;
  * @date 2020/4/5 - 19:45
  */
 public interface InstockDetailServiceDao {
-    int insertDetail(Integer goodid,double OrPrice,Integer Amount,Integer instockNweId);//添加入库明细
+    int insertDetail(Integer goodid,double OrPrice,double Amount,Integer instockNweId);//添加入库明细
 
     List<InstockDetail> selectNewDetail(Integer instockNweId);//返回入库明细
